move adapters into core directory
