+++ core/adapters/adapter.go
