+++ core/adapters/bridge.go
