+++ core/adapters/copy.go
