+++ core/adapters/doc.go
