+++ core/adapters/multiply.go
