+++ core/adapters/multiply_sgx.go
@@ -5,7 +5,7 @@ package adapters
 /*
 #cgo LDFLAGS: -L../sgx/target/ -ladapters
 #include <stdlib.h>
-#include "../sgx/libadapters/adapters.h"
+#include "../../sgx/libadapters/adapters.h"
 */
 import "C"
