+++ core/adapters/no_op.go
