+++ core/adapters/wasm.go
